# backend/openapi/saxlem-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/doctors": {
      "get": {
        "operationId": "DoctorsController_search",
        "parameters": [
          {
            "name": "specialty",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cardiology",
              "type": "string"
            }
          },
          {
            "name": "clinicId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "badiniKurdish",
                "soraniKurdish",
                "arabic",
                "english",
                "turkish"
              ]
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "female",
                "male",
                "unspecified"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          },
          {
            "name": "minimumYearsOfExperience",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 80,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 120,
              "example": "ژیان",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorPageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Filter, identifier, or pagination validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Role, capability, visibility, or tenant policy denied access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "A required security audit could not be recorded. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search the visible doctor directory",
        "tags": [
          "doctors"
        ]
      }
    },
    "/api/v1/doctors/{id}": {
      "get": {
        "operationId": "DoctorsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Filter, identifier, or pagination validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Role, capability, visibility, or tenant policy denied access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor is absent, archived, inactive for a patient, or outside the tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "A required security audit could not be recorded. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one visible doctor",
        "tags": [
          "doctors"
        ]
      }
    },
    "/api/v1/doctors/{id}/profile": {
      "get": {
        "operationId": "DoctorsController_profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorProfessionalProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Filter, identifier, or pagination validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Role, capability, visibility, or tenant policy denied access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor is not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "A required security audit could not be recorded. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the doctor professional profile",
        "tags": [
          "doctors"
        ]
      }
    },
    "/api/v1/doctors/{id}/specialties": {
      "get": {
        "operationId": "DoctorsController_specialties",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialtyResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Filter, identifier, or pagination validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Role, capability, visibility, or tenant policy denied access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor is not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "A required security audit could not be recorded. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get canonical specialties assigned to the doctor",
        "tags": [
          "doctors"
        ]
      }
    },
    "/api/v1/doctors/{id}/availability": {
      "get": {
        "operationId": "DoctorsController_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorAvailabilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Filter, identifier, or pagination validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Role, capability, visibility, or tenant policy denied access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor is not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "A required security audit could not be recorded. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get read-only availability metadata; this is not a schedule",
        "tags": [
          "doctors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saxlem API",
    "description": "Versioned Saxlem backend contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiFieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ]
      },
      "ApiErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiFieldErrorDto"
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "retryable",
          "fieldErrors"
        ]
      },
      "ApiErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "DoctorClinicResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Saxlem Medical Center"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "DoctorAvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ]
          },
          "acceptingNewPatients": {
            "type": "boolean"
          },
          "nextAvailableAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "acceptingNewPatients",
          "nextAvailableAt",
          "updatedAt"
        ]
      },
      "DoctorListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Dr. Shilan Ahmed"
          },
          "fullName": {
            "type": "string",
            "example": "Dr. Shilan Ahmed",
            "description": "Flutter-compatible alias."
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "description": "Primary specialty display name."
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "yearsOfExperience": {
            "type": "number",
            "example": 12
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "badiniKurdish",
                "soraniKurdish",
                "arabic",
                "english",
                "turkish"
              ]
            }
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public image URL, or null until an approved publication boundary exists."
          },
          "clinics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoctorClinicResponseDto"
            }
          },
          "availability": {
            "$ref": "#/components/schemas/DoctorAvailabilityResponseDto"
          }
        },
        "required": [
          "id",
          "displayName",
          "fullName",
          "specialty",
          "gender",
          "status",
          "yearsOfExperience",
          "languages",
          "profileImageUrl",
          "clinics",
          "availability"
        ]
      },
      "DoctorPageResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoctorListItemResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1,
            "maximum": 10000
          },
          "pageSize": {
            "type": "number",
            "example": 20,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "SpecialtyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "example": "cardiology"
          },
          "displayName": {
            "type": "string",
            "example": "Cardiology"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "code",
          "displayName",
          "isPrimary"
        ]
      },
      "DoctorDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Dr. Shilan Ahmed"
          },
          "fullName": {
            "type": "string",
            "example": "Dr. Shilan Ahmed",
            "description": "Flutter-compatible alias."
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "description": "Primary specialty display name."
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "yearsOfExperience": {
            "type": "number",
            "example": 12
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "badiniKurdish",
                "soraniKurdish",
                "arabic",
                "english",
                "turkish"
              ]
            }
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public image URL, or null until an approved publication boundary exists."
          },
          "clinics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoctorClinicResponseDto"
            }
          },
          "availability": {
            "$ref": "#/components/schemas/DoctorAvailabilityResponseDto"
          },
          "firstName": {
            "type": "string",
            "example": "Shilan"
          },
          "lastName": {
            "type": "string",
            "example": "Ahmed"
          },
          "licenseNumber": {
            "type": "string",
            "example": "KRI-MED-12345"
          },
          "biography": {
            "type": "string",
            "example": "A licensed cardiologist with extensive clinical experience."
          },
          "specialties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecialtyResponseDto"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "fullName",
          "specialty",
          "gender",
          "status",
          "yearsOfExperience",
          "languages",
          "profileImageUrl",
          "clinics",
          "availability",
          "firstName",
          "lastName",
          "licenseNumber",
          "biography",
          "specialties"
        ]
      },
      "DoctorProfessionalProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "licenseNumber": {
            "type": "string"
          },
          "yearsOfExperience": {
            "type": "number"
          },
          "biography": {
            "type": "string"
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "badiniKurdish",
                "soraniKurdish",
                "arabic",
                "english",
                "turkish"
              ]
            }
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true
          },
          "specialties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecialtyResponseDto"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "fullName",
          "specialty",
          "gender",
          "licenseNumber",
          "yearsOfExperience",
          "biography",
          "languages",
          "profileImageUrl",
          "specialties"
        ]
      },
...
    }
  }
}